everything adjusted, SQL fixed, sort

+++ app/src/main/java/com/example/sungwon/lifeuniverseeverything/ReviewFragment.java
@@ -58,7 +58,7 @@ public class ReviewFragment extends DialogFragment {
         mRevCat.setText(mHelper.getCategory(cursor.getInt(cursor.getColumnIndex(SQLHelper.everythingTable.COLUMN_CATEGORY_ID))));
         mRevTag.setText(mHelper.getTag(cursor.getInt(cursor.getColumnIndex(SQLHelper.everythingTable.COLUMN_TAGS_ID))));
         mRevRev.setText(cursor.getString(cursor.getColumnIndex(SQLHelper.everythingTable.COLUMN_REVIEW)));
-        mRatingRev.setRating(cursor.getInt(cursor.getColumnIndex(SQLHelper.everythingTable.COLUMN_RATINGS))/2);
+        mRatingRev.setRating(((float)cursor.getInt(cursor.getColumnIndex(SQLHelper.everythingTable.COLUMN_RATINGS)))/2);
 
         mBack.setOnClickListener(new View.OnClickListener() {
             @Override
